requirements.txt
gunicorn
uvicorn[standard]
nonebot2
nonebot-adapter-onebot>=2.4.6
aiohttp>=3.13.2,<4.0.0
aiosqlite>=0.22.0,<0.23.0
nonebot-plugin-localstore>=0.7.4
nonebot-plugin-apscheduler>=0.5.0
nonebot2[fastapi]>=2.4.4
google-genai>=1.56.0,<2.0.0
nonebot-plugin-sentry>=2.0.0
fastembed>=0.7.4,<0.8.0
psycopg>=3.3.2,<4.0.0
asyncpg>=0.31.0,<0.32.0
streamlit>=1.52.2,<2.0.0
redis>=7.1.0,<8.0.0